pyproject.toml
[tool.poetry]
name = "aws_s3_files_autosync"
version = "0.0.1"
description = "Files watchdog to autosync with AWS S3"
authors = ["johnpreston <john@compose-x.io>"]
license = "MPL-2.0"
classifiers = [
  "Natural Language :: English",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
]
readme = "README.rst"

[tool.poetry.dependencies]
python = "^3.7"
compose-x-common = "^1.0.4"
watchdog = "^2.1.9"
PyYAML = "^6.0"
jsonschema = "^4.9.0"
importlib-resources = "^5.9.0"
pytz = "^2022.1"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
isort = "^5.10.1"
flake8 = "^5.0.2"
coverage = "^6.4.2"
pytest = "^7.1.2"
pre-commit = "^2.20.0"
tbump = "^6.9.0"
mypy-boto3 = "^1.24.43"
pyclean = "^2.2.0"
twine = "^4.0.1"
pyupgrade = "^2.37.3"

[tool.poetry.scripts]
files_local_autosync = "aws_s3_files_autosync.cli:local_main"
files_s3_autosync = "aws_s3_files_autosync.cli:s3_main"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
known_first_party = "kelvin"

[tool.coverage.report]
exclude_lines = [
  "if __name__ == '__main__'"
]
[tool.coverage.run]
omit = [
  "*/cli.py"
]

[tool.tbump]
github_url = "https://github.com/compose-x/aws_s3_files_autosync"

[tool.tbump.version]
current = "0.0.1"

regex = '''
  (?P<major>\d+)
  \.
  (?P<minor>\d+)
  \.
  (?P<patch>\d+)
  (?:(?P<rc>[\S]+))?
  '''

[tool.tbump.git]
message_template = "Bump to {new_version}"
tag_template = "v{new_version}"

[[tool.tbump.file]]
src = "pyproject.toml"

[[tool.tbump.file]]
src = "aws_s3_files_autosync/__init__.py"

[[tool.tbump.before_commit]]
name = "Files format"
cmd = "make conform"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
